Rename folder to `playwright-csharp-common`
