+++ common/src/codegen/playwright-charp-common/createDotnetProjectMetadata.ts
