+++ common/src/codegen/playwright-charp-common/createOutputFileInfo.ts
