+++ common/src/codegen/playwright-charp-common/outputProjectMetadataProcessor.ts
