+++ common/src/codegen/playwright-charp-common/readme.md
@@ -1,0 +1,1 @@
+# Contains common code for Playwright Dotnet (C#) codegens: including MsTest, Nunit, xUnit
