--- common/src/codegen/playwright-charp/readme.md
@@ -1,1 +1,0 @@
-# Contains common code for playwright-csharp codegens: including MsTest, Nunit, xUnit
